# swagger.yaml
openapi: 3.0.0
info:
    title: Node.js + MongoDB API - JWT Authentication with Refresh Tokens
    version: 1.0.0
servers:
    - url: http://localhost:4000
      description: Local development server
paths:
...
    /users/refresh-token:
        post:
            summary: Use a refresh token to generate a new JWT token and a new refresh token
            description: The refresh token is sent and returned via cookies.
            operationId: refreshToken
            parameters:
                - in: cookie
                  name: refreshToken
                  description: The `refreshToken` cookie
                  schema:
                      type: string
                      example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
            responses:
                "200":
                    description: User details, a JWT access token and a new refresh token cookie
                    headers:
                        Set-Cookie:
                            description: "`refreshToken`"
                            schema:
                                type: string
                                example: refreshToken=51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550; Path=/; Expires=Tue, 16 Jun 2020 09:14:17 GMT; HttpOnly
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                        example: "5eb12e197e06a76ccdefc121"
                                    firstName:
                                        type: string
                                        example: "Jason"
                                    lastName:
                                        type: string
                                        example: "Watmore"
                                    username:
                                        type: string
                                        example: "jason"
                                    role:
                                        type: string
                                        example: "Admin"
                                    jwtToken:
                                        type: string
                                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
                "400":
                    description: The refresh token is invalid, revoked or expired
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Invalid token"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...